+++ src/router/__tests__/index.ts
@@ -10,6 +10,22 @@ describe('router42', () => {
         expect(result.payload.toState?.name).toBe('en.index');
     });
 
+    it('general, should throw if root node is not correct', async () => {
+        let create = () => {
+            new Router42({
+                name: 'auctions',
+                path: '/auctions?type',
+                children: [
+                    { name: 'index', path: '/' },
+                    { name: 'recent', path: '/recent' },
+                    { name: 'search', path: '/search' },
+                ],
+            });
+        };
+
+        expect(create).toThrow('First node in a tree should have empty name and path, e.g. `new Route({children: [...]})` or `{children: [...]}');
+    });
+
     it("general, shoulnd't allow to start twice", async () => {
         const router = createRouter();
         await router.start('/');
@@ -106,9 +122,9 @@ describe('router42', () => {
                         },
                         onEnter: async ({ asyncResult, passthrough }) => {
                             expect(asyncResult).toBe('user.review(async)');
-                            await new Promise((resolve) => {
+                            return await new Promise((resolve) => {
                                 setTimeout(() => {
-                                    resolve('user.review');
+                                    resolve();
                                 }, 1000);
                             });
                         },
@@ -131,9 +147,236 @@ describe('router42', () => {
         let wait1 = router.navigate('user.orders', { id: 1 });
         let wait2 = router.navigate('user.review', { page: 1 });
         const results = await Promise.all([wait1, wait2]);
-        expect(results[0].type === 'error').toBeTruthy();
-        expect(results[0].payload.error?.code === 'TRANSITION_CANCELLED').toBeTruthy();
-        expect(results[1].type === 'success').toBeTruthy();
+        expect(results[0].type).toBe('error');
+        expect(results[0].payload.error?.code).toBe(errorCodes.TRANSITION_CANCELLED);
+        expect((results[1] as any).type).toBe('success');
+    });
+
+    it('transition, should cancel right after onEnter func', async () => {
+        const router = new Router42([
+            {
+                name: 'orders',
+                path: '/orders',
+                children: [
+                    {
+                        name: 'index',
+                        path: '/',
+                    },
+                    {
+                        name: 'top',
+                        path: '/top/:id',
+                        onEnter: async () => {
+                            return await new Promise((resolve) => {
+                                setTimeout(() => {
+                                    resolve();
+                                }, 1000);
+                            });
+                        },
+                    },
+                    { name: 'statistics', path: '/statistics?type' },
+                    { name: 'drop', path: '/drop' },
+                ],
+            },
+        ]);
+
+        await router.start('/orders');
+        let firstNavigarion = router.navigate('orders.top', { id: 1 });
+        let simulatedDelay = new Promise((resolve) => {
+            setTimeout(() => {
+                resolve(router.navigate('orders.drop'));
+            }, 250);
+        });
+
+        const results = await Promise.all([firstNavigarion, simulatedDelay]);
+        expect(results[0].type).toBe('error');
+        expect(results[0].payload.error?.code).toBe(errorCodes.TRANSITION_CANCELLED);
+        expect((results[1] as any).type).toBe('success');
+    });
+
+    it('transition, default params', async () => {
+        const router = new Router42([
+            { name: 'index', path: '/' },
+            { name: 'section', path: '/:section', defaultParams: { section: 'kek', q: 'q' } },
+        ]);
+
+        await router.start('/');
+        let result = await router.navigate('section');
+        expect(result.payload.toState?.params).toEqual({ section: 'kek', q: 'q' });
+    });
+
+    it('transition, redirect', async () => {
+        const router = new Router42(
+            [
+                { name: 'index', path: '/' },
+                { name: 'auth', path: '/auth' },
+                {
+                    name: 'protected',
+                    path: '/protected',
+                    onEnter: ({ dependencies }) => {
+                        if (!dependencies!.authorized) {
+                            throw new NavigationError({
+                                code: errorCodes.TRANSITION_REDIRECTED,
+                                redirect: { name: 'auth', params: { return_to: 'protected' } },
+                            });
+                        }
+                    },
+                },
+            ],
+            {},
+            { authorized: false }
+        );
+
+        await router.start('/');
+        let result = await router.navigate('protected');
+        expect(result.payload.toState?.name).toBe('auth');
+        expect(result.payload.toState?.params).toEqual({ return_to: 'protected' });
+    });
+
+    it('transition, pinpointed params in path are matter', async () => {
+        let onEnter = jest.fn();
+        const router = new Router42([
+            {
+                name: 'auctions',
+                path: '/auctions?type',
+                onEnter: onEnter,
+                children: [
+                    { name: 'index', path: '/' },
+                    { name: 'recent', path: '/recent' },
+                    { name: 'search', path: '/search' },
+                ],
+            },
+        ]);
+
+        let result = await router.start('/auctions');
+        result = await router.navigate('auctions.search', { type: 'lich', weapon: 'hek' });
+        result = await router.navigate('auctions.search', { type: 'riven', weapon: 'tonkor' });
+        expect(result.payload.toActivate?.length).toBe(2);
+        expect(result.payload.toDeactivate?.length).toBe(2);
+        expect(result.payload.toActivate![0].name).toBe('auctions');
+        expect(result.payload.toDeactivate![0].name).toBe('auctions');
+        expect(onEnter.mock.calls.length).toBe(3);
+        // omit one param, should detect and reactivate `auctions` node anyway
+        result = await router.navigate('auctions.search', { weapon: 'tonkor' });
+        expect(result.payload.toActivate?.length).toBe(2);
+        expect(result.payload.toDeactivate?.length).toBe(2);
+        expect(result.payload.toActivate![0].name).toBe('auctions');
+        expect(result.payload.toDeactivate![0].name).toBe('auctions');
+    });
+
+    it('transition, catch unknown error', async () => {
+        const router = new Router42([
+            { name: 'index', path: '/' },
+            {
+                name: 'test',
+                path: '/test',
+                onEnter: () => {
+                    throw new Error('Unexpected error');
+                },
+            },
+        ]);
+
+        let cb = jest.fn();
+
+        router.addEventListener(events.TRANSITION_UNKNOWN_ERROR, cb);
+
+        await router.start('/');
+        let result = await router.navigate('test');
+        expect(result.type).toBe('error');
+
+        expect(cb.mock.calls.length).toBe(1);
+        expect(cb.mock.calls[0][0]['error']['message']).toBe('Unexpected error');
+    });
+
+    it('transition, navigation should work if states are the same and force == true', async () => {
+        const router = createRouter();
+
+        await router.start('/');
+        await router.navigate('en.profile.index', { name: 'KycKyc' });
+        let result = await router.navigate('en.profile.index', { name: 'KycKyc' }, { force: true });
+        expect(result.type).toBe('success');
+    });
+
+    it('transition, navigation should not work if states are the same and force == false', async () => {
+        const router = createRouter();
+
+        await router.start('/');
+        await router.navigate('en.profile.index', { name: 'KycKyc' });
+        let result = await router.navigate('en.profile.index', { name: 'KycKyc' });
+        expect(result.type).toBe('error');
+        expect(result.payload.error?.code).toBe(errorCodes.SAME_STATES);
+    });
+
+    it('transition, reload should work', async () => {
+        const langEnter = jest.fn();
+        const indexEnter = jest.fn();
+        const mainNodes = [
+            {
+                name: 'auctions',
+                path: '/auctions?type',
+                children: [
+                    { name: 'index', path: '/', onEnter: indexEnter },
+                    { name: 'recent', path: '/recent' },
+                    { name: 'search', path: '/search' },
+                ],
+            },
+        ];
+
+        const router = new Router42([
+            { name: 'en', path: '/', children: mainNodes, onEnter: langEnter },
+            { name: 'ru', path: '/ru', children: mainNodes, onEnter: langEnter },
+            { name: 'ko', path: '/ko', children: mainNodes, onEnter: langEnter },
+        ]);
+
+        await router.start('/auctions');
+        let result = await router.navigate('*.auctions.index', {});
+        expect(result.payload.error?.code).toBe(errorCodes.SAME_STATES);
+        result = await router.navigate('*.auctions.index', {}, { reload: true });
+        expect(result.payload.toActivate?.length).toBe(3);
+        expect(langEnter.mock.calls.length).toBe(2);
+        expect(indexEnter.mock.calls.length).toBe(2);
+    });
+
+    it('transition, reload, with nodes that ignore reload calls, should work correctly', async () => {
+        const langEnter = jest.fn();
+        const indexEnter = jest.fn();
+        const mainNodes = [
+            {
+                name: 'auctions',
+                path: '/auctions?type',
+                children: [
+                    { name: 'index', path: '/', onEnter: indexEnter },
+                    { name: 'recent', path: '/recent' },
+                    { name: 'search', path: '/search' },
+                ],
+            },
+        ];
+
+        const router = new Router42([
+            { name: 'en', path: '/', children: mainNodes, ignoreReloadCall: true, onEnter: langEnter },
+            { name: 'ru', path: '/ru', children: mainNodes, ignoreReloadCall: true, onEnter: langEnter },
+            { name: 'ko', path: '/ko', children: mainNodes, ignoreReloadCall: true, onEnter: langEnter },
+        ]);
+
+        await router.start('/auctions');
+        let result = await router.navigate('*.auctions.index', {});
+        expect(result.payload.error?.code).toBe(errorCodes.SAME_STATES);
+        result = await router.navigate('*.auctions.index', {}, { reload: true });
+        expect(result.payload.toActivate?.length).toBe(2);
+        expect(langEnter.mock.calls.length).toBe(1);
+        expect(indexEnter.mock.calls.length).toBe(2);
+    });
+
+    it('transition, route not found, no fallbacks are defined', async () => {
+        const router = new Router42([
+            { name: 'index', path: '/' },
+            { name: 'one', path: '/one' },
+            { name: 'two', path: '/two' },
+        ]);
+
+        await router.start('/');
+        let result = await router.navigate('nowhere');
+        expect(result.type).toBe('error');
+        expect(result.payload.error?.code).toBe(errorCodes.ROUTE_NOT_FOUND);
     });
 
     it('404, incorrect node name', async () => {
@@ -218,7 +461,7 @@ describe('router42', () => {
         expect(result.payload.toState?.name).toBe('ko.index');
     });
 
-    it('status, route is active, full path', async () => {
+    it('status, route is active, exact path', async () => {
         const router = createRouter();
         await router.start('/');
         await router.navigate('ru.profile.auctions', { name: 'KycKyc', kek: 'pek' });
@@ -227,13 +470,12 @@ describe('router42', () => {
         expect(router.isActive('ru.profile.auctions', { name: 'KycKyc' }, true, false)).toBe(false);
     });
 
-    it('status, route is active, some child', async () => {
+    it('status, route is active, checking some child node', async () => {
         const router = createRouter();
         await router.start('/');
         await router.navigate('ru.profile.auctions', { name: 'KycKyc', kek: 'pek' });
-        expect(router.isActive('ru.profile.auctions', { name: 'KycKyc' })).toBe(true);
-        expect(router.isActive('ru.profile.auctions', { name: 'KycKyc', kek: 'pek' })).toBe(true);
-        expect(router.isActive('ru.profile.auctions', { name: 'KycKyc' }, true, false)).toBe(false);
+        expect(router.isActive('ru.profile', { name: 'KycKyc' }, false)).toBe(true);
+        expect(router.isActive('ru.profile', { name: 'KycKyc' })).toBe(false);
     });
 
     it('events, start & stop', async () => {